requirements.txt
aniso8601==9.0.1
bidict==0.22.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.3
Flask-WTF==1.0.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mongoengine==0.24.2
oauthlib==3.2.0
osrparse==6.0.1
ossapi==2.3.0
pymongo==4.2.0
python-dotenv==0.20.0
python-engineio==4.3.3
python-socketio==5.7.1
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
typing-utils==0.1.0
urllib3==1.26.10
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.1
